pyproject.toml
[project]
name = "galnet-chronicle"
version = "0.1.0"
description = "The GalNet Chronicle — an enriched archive of Elite: Dangerous GalNet articles"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
]
